+++ R/pycnogrid-package.R
@@ -1,0 +1,4 @@
+#' pycnogrid
+#'
+#' @importFrom rlang .data :=
+"_PACKAGE"